Insert confession
If a spirit confesses the following: the son of man came in the flesh and died and rose from the dead, then it is of the Anointed One. If it says otherwise. then it is of the opposition to the Anointed One, the Anti-Anointed, which you have heard will enter into the world, and is indeed here. Why is this? Not all spirits that come to you are of the Anointed One, who is with G!D and of G!D - for there are many false prophets that have come into the world. So, the test of whether is spirit is of the One who was given Ointment, the Anointed One, or the Anti-Anointed, is the following: The spirit that is of the Anointed One confesses that the son of man came in the flesh, and died, and rose from the dead. (Amen - which means, 'So it is' or 'so be it'; it is uncertain)