Re
The servant is not above the master, and the student is not above the teacher, and the messenger is not above the One that sends that message, but a fully-trained student will be like the teacher. Truly, truly, I tell you. - Jesus, son of Mary, whom G!D enabled to conceive when no man touched her with even a single hair