With G!D's help, I have an index so that others can search easier!
Praise be to G!D, who cherishes and sustains the Worlds. Wow. Great an marvellous are
Your deeds - fabulous and wonderful and miraculous, like a rainbow after a rain-storm!
O Almighty One, King of the ages, King of the nations, just and true are Your ways and judgments.
Who will not fear You? For You are the only holy one. All will come and worship You, for
your righteous acts are revealed. You who were, and you who are: the eternal G!D, the one and only, my one and only, who
has long life. And sang the song of Moses, who is servant of G!D, and of the Lamb