With G!D's help, I have communication!
It's not californication,
it's communication with the help
of G!D who grants boats passage across
oceans known and unknown even. G!D
who created sky for ceiling for Man, and
earth for bed for Man (me). Thank you, Compassionate
and Most Merciful Cherisher and Sustainer of the Worlds,
whose ways and judgments are true and just indeed. Those
who have ears, let them hear. The Son of Man is coming.
The kingdom of G!D is near!