Thank You, Compassionate and Most Merciful One
Thank You for giving me what I have, even me.
Please help me, through the one who has the name Jesus, to love 
You more, and do what I cannot, which is to love You with unknown parts
of me, and known parts of me, and awakening parts of me. Second to that
 (which is first and foremost)
please help me to do what I cannot, which is to keep Your commandments 
and love thy neighbour as myself.
To G!D be the glory for-ever and ever. Amen. Praise be to G!D.