Praise be to G!D
Jesus took bread, gave thanks to Father who is blessed indeed, the Ruler of the Universe, whose name is Hallowed. Then he gave to the Twelve (apostles), and said, 'This is my body, which is for you. Eat it, and do this
frequently in remembrance (or loving memory) of Me.'