By Your will, Daddy, Papa, Pa, not by my will, let this be done
The Son of Man was o, he was sorrowful to point of death.
He went up a mountain to pray. 'If this cup, I must drink from,
let it be by your will, Father, not by my will, that I drink from
this cup.' This was the second time that Jesus, who called himself
the Son of Man, prayed. How? He left his disciples. Or, he went away
from his disciples. (Matthew 26:42)
http://biblehub.com/matthew/26-42.htm