A servant or bond-servant of the Lordmust be patient
A servant or bond-servant of the Lord must be patient,
able to teach.
Lo! Those who believe (in that which is revealed unto thee, Muhammad), and those who are Jews, and Christians, and Sabaeans - whoever believeth in Allah and the Last Day and doeth right - surely their reward is with their Lord, and there shall no fear come upon them neither shall they grieve.
(Quran. chapter 2, verse 62.)
 http://corpus.quran.com/translation.jsp?chapter=2&verse=62

Children of Israel, recall when We made a covenant with you, raised Mount Tur (Sinai) above you and told you to receive earnestly what We had given to you and bear it in mind so that you would protect yourselves against evil.
or,
And (O Children of Israel, remember) when We took your covenant and We raised above you the Mount (saying): "Hold fast to that which We have given you, and remember that which is therein so that you may become Al-Muttaqun (the pious - see V.2:2).
or,
And when We took compact with you, and raised above you the Mount: 'Take forcefully what We have given you, and remember what is in it; haply you shall be god-fearing.
(Scriptures that some
call the Quran. chapter 2. verse 63).
http://corpus.quran.com/translation.jsp?chapter=2&verse=63