With G!D's help, I have a step to clarity yet again!
The re is story of martial artist. Reached the peak
in man-made measure of success. You know, 
ranking among his peers. Colour of belt he wear,
so on, and so forth. Then when he practise martial arts,
he wear the belt that is reserved for beginners to the discipline,
instead of the belt that is reserved for the top.