Go into all nations
, baptising them into Dad, Son, and G!D's spirit,
teaching them to obey everythin' I have taught you,
and commanded you, and to obey the commandments
(you know the commandments). An dsurly, I will
always be with you, for-ever and always, until
the very end of the world, and the passing away of
this generation, and teh very end of the age, even the last day.
Go into all nations, baptising them into the Father, the Son,
and the Spirit of G!D, who hath no begotten son. Amen.
Through the One who has the Name, Jesus. Christ Jesus.
(as written in the Quran, and the Bible, and from G!D-knows-what, and
I do not know. Yes, G!D knows, and I do not know.)