Help me to love You, G!D
Help me to love You, G!D, through Christ Jesus, first and fore-most.
Not love for worldly wine or champagne or money, oh Lord of lords
King of kings, but love for You.
Help me to do what I cannot, but by Your Almighty
power, I can: to love You with parts of me that are known
to me, and unknown to me. Please You, first and fore-most,
help me to do so, too. Not how to please husband that is of
the world, or wife who is of the world, or any who are
of the world, but You. Please help me to please You, and only
You, through Christ Jesus. 

Jesus, son of Mary, who calls himself
the Son of Man, came in the flesh - 
he had bones and flesh - and rose
from the dead. As surely as the LORD
lives, as surely as G!D lives, I confess and
testify this: that I have witnessed the rising
of the Son of Man from the dead after he
died, and that he has come in the flesh,
eating and drinking, and having bones 
and flesh, for a spirit has no flesh, and no bones.

Thank you, blessed Father in Heaven, for gardens
of bliss, the success, the objective of life, the succor,
gardens under which rivers gush, and streams flow,
and wherein spouses relax and recline, and delight 
each other, for the good pleasure of G!D. Yes, the supreme
felicity, the success, the succor, the objective of life. My portion
in the LORD G!D, all the rest of my meaningless life-days, this is,
as G!D hath guided me to take from Him. So blissful. And even
a blissful oblivion to the mundane affairs of the world, You give me,
in lovingkindness, to a little one, and a worm, such as myself.
Daddy, I love You, Papa, who says:
'who says fathers will eat sour grapes, and children's
teeth are set on edge? 
from now on, all those who eat sour grapes,
their own teeth are set on edge.' (as written in Scriptures,
which you can check in what some call the Bible.) 

:-)

^_^