Praise be to G!D, who helped me, even me
With G!D's help, I have a step to clarity, and another step, and still
other steps that G!D knows, and I do not know. Yes, steps. Steps
that are unknown to me. O, good. O, taste and see that G!D is good!
Blessed are those who take refuge in Him!