The kingdom belongs to G!D
All my base is belong to You, oh Almighty One who cherishes
and sustains the worlds: the Sovereign over the Last Day,
who created all. I'm feeling good and cheerful, O blessed Father 
who loves justice and detests wickedness, and is severe
and swift and terrible in punishment that is eternal or other-wise
a cause of agony and torment.

Thank You, Compassionate and Most Merciful One :-) You are true in word and deed,
and just in ways and judgments, King of kings, Lord of lords 

Who will not fear You? For You are the only holy one. All will come,
 and worship You, for Your righteous acts are revealed. You who was, 
and who is: the Everlasting G!D, God of Abraham and the prophets

Sang the song of Moses, who is servant of G!D, and of the Lamb, who
is worthy of glory and honour and power and kingdom and praise.