He must increase; I must decrease
The bride is the bridegroom's.
The bride belongs to the bridegroom.
All base of bride is belong to bridegroom.
But the friend of the bridegroom hears
his voice, and rejoices. In this my joy is made
full. He must increase; I must decrease.
(LORD G!D be glorified through my deeds;
making wish through Christ Jesus)
(determined-little-one meme-picture) 
:-)