With G!D's help, I have convenience to offer Man!
It is not by my righteousness that these
have come to me. It is because of
the wickedness of those who have
done wickedness in the sight of the 
LORD G!D - who is the All-knowing,
All-hearing, and All-seeing - and
He has driven them out from before me.
Yes, the Ruler of the Universe, who is the
Compassionate and Most Merciful One,
has driven nations that are mightier and
greater than I  - or used to be mightier
and greater than I - out from before me.
And that is why I have what I have, and can
do what I can do, and is becoming what I am
becoming, and am what I am, and is being what
I am being.
Praise be to the just and true king of the ages, the
king of the nations, the Most Merciful and Compassionate
Lord of lords, King of kings, to whom I belong. (and proudly
I say so; that I do belong to the LORD G!D, in sweet
surrender)
And the LORD G!D is not unmindful of your deeds.