To G!D who loves justice be glory for-ever n ever
Let not their marvels of steel
dazzle you. Nothing that is made is made
without　Ｇ！Ｄ

Do not be angry. 
Why? Because anger is in the bosom of the fool.
Anger is like consuming a cockroach-bait
(that is, what 'modern' people promise will kill cockroaches -
quite publicly marketed to the extent that it is highly
likely you will find one, near-by you, on sale)
anger is like consuming poison that supposedly
kills cockroaches, and when the cockroaches
are running around in a forest, you expect that
when you swallow and vigorously digest cockroach bait,
they will die instead of you.
So, do not be angry, for anger is in the bosom of the fool.