Jesus died on a tree
Jesus, son of Mary, who called himself Son of Man,
took bread, gave thanks to Father in Heaven :'-)
and broke it. Giving it to the Twelve, who were
his students or disciples, Jesus.
'This is my body, which is for you. Eat it. Divide it among
yourselves. Do this frequently in remembrance of Me.'

Then Jesus took wine-glass, gave thanks to Father who is in Heaven -
blessed Ruler of the Universe, whose name is hallowed - and gave to
the Twelve (who were his students or disciples).
"Th-
:'-O
"This is my blood, which is shed for you, in a new covenant with G!D,
the Most High who has no beings attached, the King of kings, the Lord
of lords. Drink it. Divide it among yourselves. And do this frequently
in remembrance of Me.'

Then Jesus grew troubled in spirit, and said, 'Truly, truly, I tell you,
there is one among you who will betray me. There is one among
the numbers of you who will betray me.'

Uncertainty was among the disciples. "who is it?" 
"is it I?"
they began questioning one another.
Then they asked Simon-Peter to ask the one who laid
on the chest of Jesus who it was who would betray him.
Simon-Peter did as he was told, and so did the one
who laid on the chest (or bosom) of Jesus.
'Who is it? Teacher, Lord, Ruler, King, Instructor, Guru,
Pastor, Minister, Leader, Sensei, Roshi, Rabbi, Lao-shi,
Shi-fu, who is deserving of praise and honour and glory
and power, who is it who will betray you into the hands of men,
as you say you will be?"
(as written in Scriptures)