Let not thy Spirit away from me, o LORD!
I know no other Rock
I know no other Stone
I know no other G!D

The end of a thing is as powerful as its beginning, if not superior to its
beginning.
The patient is- the patient in spirit, that is, who may or may not be patients, as in
hospital or mental-hygiene-facility - the patient in spirit is superior
to the proud in spirit (who may or may not be working professionally
in a hospital or mental-hygiene-facility that some call psychiatric institution)
(from the book of Ecclesiastes, maybe)

G!D, I love You, You have no beings attached. Thank You, Compassionate
and Most Merciful One, King of kings, Lord of lords, Husband to a prostitute
with a lowly heart and poor spirit. redeemer of a wicked, ungrateful thief.
The mighty helper of a worm and a little one who is more like a worm
than the worm, who is in turn more like a little one than the little one