Daddy, I love You so.
Help me to love the other, who may or may not be unknown to me, as You have
loved me, even me.
through Christ Jesus, whom I am not above. 
Amen.
Praise be to G!D Most High, who cherishes and sustains the worlds.
G!D knows how to help them, and I do not...
Ah, those fools who associate themselves with pursuit of world-desires, all of which are fleeting
and pass away in their present form, together with everything of the world
and the world itself - nothing but foolishness they are doing, a certificate
that does not guarantee employment, money that does not guarantee happiness or
even acceptance at a retail shop (try rupiah at a store that's not in the Indonesian
archipelago), sex that leaves you hungry and thirsty for more even after you've had it,
o, no matter what kinky things you do in private or in public...
'What do we do?' they asked him. (from book entitled 'Praise to a formless god')
He said to them: 'You are trapped! You are trapped in sexual play with women
who lust after things of the world and the world itself - jewels, cute boys, perfume,
hair colour, fancy restaurants, tasteful appearances in nude photography
(that may or may not appear in public),
well-appointed rooms in boutique hotels - trapped in play with them, in sexual matters,
you are!'