G!D, help me to resist the tempter A.K.A. Belial
do not err, brethren, as written in Scriptures,
every good gift, and every perfect gift,
is from the blessed Father, who is in Heaven,
and unseen before men, (or oh- you get the drift,
women, and those who are of the world),
with whom there is no passing away, no fleeting-ness,
no turning of shadow - for He is called
the Everlasting god
through Christ Jesus

"Heaven forbid that you should die, Teacher, Ruler over us,
Lord, our Owner," said Simon-Peter to Jesus, who had made
it plain to his students, or disciples, that he was going to die.
O, immediately Jesus, who heard what Simon-Peter was saying,
 turned to Simon-Peter, and said, "Behind me, get behind me, please,
Satan, also known as Belial or Shaitan - tempter who calls himself
Angel of Light."

"What you say," Jesus said, -oh,

"Get behind me, please, Satan. You are a stumbling block to Me!

"What you say comes from the mind of Man, which is not
above the Mind of G!D. The mind of G!D is above the mind
of Man, and the ways of G!D are above the ways of Man,
and the thoughts of G!D are above the ways of Man"

(So Simon-Peter did not get condemned to eternal Fire,
for speaking his mind)
Oh, how patient and understanding our Lord and Teacher,
whom I am not above, is! Tender-hearted, too!