Praise be to G!D who is Cherisher of the worlds
To G!D who cherishes and delights me, and sustains
the worlds, even me, thank You!
:-)
^-^
(so happy)
I praise You, for You have shown
lovingkindness 
:-)
or, kindness, that is given out of love,
unconditionally, without expectation of
return, that is, to me, even me, a little one,
and a worm, not to mention a fool
and thief (or the one men call "ungrateful")
And may- I hasten to count myself among the wicked
may the love of G!D be with you.
(Amen. Through Christ Jesus.)
Amen. Through Christ Jesus.
Let it be so, O Father in Heaven!