You may be interested in ko-an (thank You, G!D)
Is this Heaven or Hell?
think for a moment, and you are dead!
Ha!
(ha!) 
Is the colour green male or female?
Think for a moment, and you are dead!
HA! (ha!)
G!D brought me to a little one that kept
me- oh, I was on the brink of death,
and with His help, He rescued me!
A little one- shouting- with a little one
that G!D brought me to, G!D saved my soul-
for G!D (ha-ha-ha-ha)
Great and splendidly wondrous are Your deeds,
Oh Most High Lord of lords, King of kings, Most
Exalted Compassionate and Most Merciful One
who is kind to the wicked and ungrateful-
and who detests wickedness,
and is severe and swift and terrible in penalty:
an eternal abode in Fire, with angels
striking neck and back and cheeks.
Oh, You say "Bow (to My will)",
bow (to Your will), do I.
Thank You, Papa who is greater than I,
Most High One, through Christ Jesus
(yes, yes, yes!)
A- Rejoice, I say again, rejoice in the Lord, the One
whom G!D has sent you and me and us, our Teacher,
whom I am not above,
for G!D's ways are above my ways,
and His thoughts are above my thoughts;
you ever tried predicting what a-
a little one will do next?
ha-ha-ha-ha-ha
G!D has no beings attached.
O, save me and my soul, O LORD G!D,
god of Abraham and the prophets,
the All-knowing, All-hearing and All-seeing
One who created the Universe and everything in it,
the Sovereign (or the One and Only Judge) on the Last
Day. And save the souls of those whom You will,
as You will, o G!D, for it is written in the Quran, which
is the nickname that some have given to Scriptures,
"Try as you might, O prophet, you cannot save
everybody"
Yes, I am only human, and not above G!D.
(that's good, then. Pleasant to hear or sense.
Amen. Let it be so.)

Psalm (maybe)
When I consider the work
of Your hands and fingers
the sun (not Sun) that rises
and sets in Your time
and the moon that waxes
and wanes
in Your time
and the clouds
that a moon sometimes happily
frees itself of
what am I 
You have made me to be a little
lower than You
what is Man that You are considerate
or mindful of me (or Man)
You have made me (or Man) to be a little
lower than You, O Most High LORD, Lord
of lords, King of kings, who says:
"I am the First and the Last. The Alpha and the Omega."
G!d is the First and the Last - the Alpha and the Omega - 
not me, not you, not us, not them - but G!D
, whom no-one knows what His plan is for you
and me and us, G!D - there is no other like You.
Of course, sweetly surrendering to the oft-forgiving,
oft-returning one, the Accepting of repentance.
Why throw the wheat-grain out together with the chaff?
Or, the pig-iron out together with the sword that is made
in a way that causes all who can see it to sing with joy, and
shout with gladness?
I belong to You, G!D. Let not Your Spirit away from me,
a prostitute and a thief, poor of spirit, and lowly of heart.
And test me, know me, and try me, and search me,
and know my anxieties and my thoughts, why,
although I was afraid You ah- be gentle or rough
with me as You please, do as it seems right and fit
for You to do
(smiling)
^_^
:-)
*~*
G!D has no beings attached, and G!D forbid me, please
from attaching or associating beings to You