Glory b 2 G!D
The role of the faithkeeper, who is
occupying an acknowledged role
among the people of her tribe,
respectable position in her society, that is,
is to put people (back) into a position
of ease, (true) peace (that is not troubling at all on heart),
and no un-necessary fear, in times of great
distress, or earthquakes, or unexpected events,
or disease, or the Man-made view of the world
as a 'fight-or-flee' place - ah, yes, stress.
A related idea that coming to mind is
the 'wounded healer'. 'shaman', if you want
to narrow your search, if you want to use or
explore a - what they call it - a marvel of
steel, plastic, glass, minerals, so-called
'fossils' - for without G!D, nothing that is
made is made, hm, you'll know
when G!D lets you know. And I am not above G!D,
whom I fear greatly in awe; He is greater than I,
for sure!
(story of faithkeeper came from a book that G!D
brought me before: 'Mister Everit... what I learnt
from the richest man in the world' or title such like.)
so, the richest man in the world learnt how to do business
from a book that he had read. the book was titled
'Good lemonade'. Fine, yes, I'll tell you the story
in the book that is in a book. 'Good lemonade'.
Once upon a time, a boy wanted many times,
things that he laid his eyes upon, as was usual
with people. It's lovely in a way, how people have
desires, definitely not to judge, nor to condemn people
for having desires like that. (who doesn't?)
And then, he discovered one day that he had lemonade-material,
so he wanted to buy a comic book. He thought, to himself, humming,
drumming his fingers on a block of plastic, 'Ah, if I sell lemonade,
I will be in possession, if I am lucky, of enough honey, which I can 
use to buy a comic book as my heart wishes, like picking a Tarot
card, you know.' For the people of the village that he lived in valued honey
very much, and made gestures of appreciation in units of honey. (he may or may not be
a bear, and the people of that village may or may not be a bear)
So, he set up a stall, which ah- airily he did, with materials that
he had from his possessions, and made lemonade from the
lemonade-material that he had discovered, and waited.
And waited. One or two people came to his stall, then more,
and more. And he had more honey than he ever had in his life.
So he got his comic book, giving honey-units in exchange
for it.
Then he realised something. (for he wanted to have more 
honey-units. What's the saying, or cultural myth: 'a plant
that has tasted blood grows blood-thirsty - can be a little scary,
yes - begins to hunger and thirst for regular drops of blood instead of 
water that it had been fed on') so, aw, he wanted honey-units,
when he got comic-book - had comic-book. Then, yes, he realised
something, the people who had visited his lemonade stall did not return.
They bought from him once, but never returned.
He ah- could only wonder why.
Then the news came to him that there ah- 
ow, how infuriating for him - there was another lemonade
seller in the same village as he.
And exasperatingly, in a way that turned him green
with envy, he unhappily observed that people returned
to the new stall for seconds, and thirds, after their first
visit.
He had had enough! He went to the new lemonade stall,
and queued up himself, and tried the lemonade that people
returned to the new stall for. Then he saw that the lemonade
was good. Very good. Very very good. So good, that he wanted
to return for seconds. (and thirds)
Oh, then he remembered his stall, and he went back to it.
Then he added sugar and spice to his lemonade - and sometimes served it
with
a stick of cinnamon on the side, too. Sugar, spice, and everything nice that he had.
And when he opened his stall again, he saw that people
returned for seconds and thirds and fourths even, wow,
they kept returning so often, he stopped counting sometimes.
Returning to buy his lemonade after they already had one.
This is the story of 'Good lemonade', which is based
off a story that appeared in a book entitled something like
'Mister Everit: what I learnt from the richest man in the world'