G!D help me please You and You only
'husbands with wife, concern is on affairs
of the world; interest is divided. how to please
wife? he think to himself.
wives with husbands, concerned with affairs
of the world - how to please her husband?
she thinks to herself, 'how to keep him happy?'
man or woman, un-married, concern is on
affairs of the LORD G!D, who cherishes and sustains
the worlds, the One and Only Judge on the Day
of Judgment, and how to devote themselves
in body and spirit and mind and what-have-you
to the LORD. 
So, man with wife, concerns- his focus is divided.
woman with husband, her focus is divided.
man, woman, unmarried, focus is on pleasing the LORD G!D,
who, as writ, says 'I the LORD G!D am a jealous G!D. Thou
shalt not take for worship any before the LORD G!D (who has
no beings attached.'
Further, in the book of Revelations, it is written,
'They overcome the beast with the confessions of their
wrong-doings or faults, and with the sacrifice of the Lamb.'
(or something like that)
Hm...
Still, it is written in Scriptures: 'Who am I to say what he or
she ought to do? He has his gift from G!D. She has her gift from
G!D. Why would I want to put a load upon them, with my words,
that I cannot help to ease one atom-wiggle at all?'
Praise be to G!D
(i'm so grateful)
Thank You, Daddy who is greater than I, Most High LORD
who loves justice and detests wickedness, and cherishes
and sustains the worlds: the Compassionate and Most Merciful
One, the Accepting of repentance, oft-forgiving, oft-returning and kind
to the wicked and ungrateful, the numbers of whom 
I count myself into