Anna the prophetess thanked G!D
Anna, a very old woman,
a señorita
(a saying from the Chinese language: '
the old ginger is the spicy ginger, the ginger
that has sufficient kick to taste-buds'
ha-ha-ha-ha (laughing)) 
señorita, señor, senile, derived from the word
(Latin maybe) 'sen', which means 'old'. related
note, this is, since G!D has given me such an
argument - of course, this is a supposition,
a claim, which may or may not be true - false knowledge,
as valid an argument nonetheless as the argument that
follows: 'the moon is blue'
there is book titled 'everything is an argument' if you
are interested, but beginning godless chatter- so-
Anna, señorita, seventy-eight years
and eighty-seven days old, widowed, who did
nothing except fasting and praying to G!D, and praising
G!D of course, in temple, took the new-born
(the Anointed One, that is) into her arms, with
the permission of those were holding him, 
and thanked G!D, and sang praises to G!D.
And returned the Anointed One, new-born,
to the arms of those who had held him.
And Joseph (the man) and Mary, mother of
the Anointed One, who had conceived with help of
G!D, and help of no man, returned with the Anointed
One to their home-town.